change int to double in appMenu

+++ src/lt/vtmc/ems/AppMenu.java
@@ -18,13 +18,13 @@ public class AppMenu {
     private void arithmeticCalculateMenu() throws Exception {
         // TODO
         System.out.print("First number: ");
-        int x = Integer.parseInt(input.next());
+        double x = Double.parseDouble(input.next());
         System.out.print("Second number: ");
-        int y = Integer.parseInt(input.next());
+        double y = Double.parseDouble(input.next());
         System.out.print("Operator: ");
         String op = input.next();
 
-        int z = 0;
+        double z = 0;
 
         if (op.equals("+")) {
             z = x + y;
